test(intproxy): add mirrord ui session monitor contract tests
Lock down the wire contract between mirrord-intproxy's per-session
HTTP API and mirrord-session-monitor-client (used by 'mirrord ui')
ahead of the eventual Windows transport port.

- Stub mirrord ui on Windows so the CLI builds while transport work
  is pending: split cli/src/ui.rs into a small dispatcher and a unix
  ui_impl.rs; on Windows ui_command returns CliError::UiError. The
  session-monitor-client dep moves to cfg(unix) in cli/Cargo.toml,
  with extern-crate suppressors in main.rs for the Windows build.
- Refactor start_api_server to take an explicit sessions_dir param so
  tests can run against a tempdir without racing on $HOME. The cli
  caller now passes home_dir().join('.mirrord/sessions').
- Add 9 cfg(unix) integration tests in mirrord/intproxy/tests/
  session_monitor_round_trip.rs, each spinning up its own server in
  a tempdir so they parallelize:
  * /info round-trip
  * /health endpoint
  * SSE /events shape for LayerConnected
  * Layer connect/disconnect updater
  * Duplicate-pid dedup
  * PortSubscription upsert by port
  * Path-traversal session_id rejection
  * Sessions dir 0o700 / socket 0o600 perms
  * /kill graceful shutdown + socket cleanup
- TEMP(ci): force the test/build jobs to fan out on workflow_dispatch
  events too, so the fork's manual dispatch can run the full matrix.
  Revert this stanza in ci.yaml before merge upstream.
