+++ changelog.d/+windows-mirrord-ui.added.md
@@ -1,0 +1,1 @@
+Added support for `mirrord ui` on Windows.
